Changed homogenization routine
Changed homogenization routine in order to account for fibers inside RVE.  Tested against uniaxial stretch cases (compressible Hex8 and incompressible Tet10) and compared with Abaqus. Seems to be working.

OBS.1: homogenization procedure to be confirmed at meeting with advisors
OBS.2: there is a bug when using the Taylor multiscale boundary condition; it takes the penultimate load step as the last! Confirmed with Thiago.